avoid running out of memory using `write file` adapter
the fix for issue #6 is currently to set a write limit so that we
simply limit writing out no more than 100K JSON points in the currently
supported writing format of JSON.

while fixing this issue I've also moved the writes to happen
asynchronously using the scheduler to schedule the syncing of data to
and from disk.

future work could be to include streamable writing formats such as CSV,
or JSONL (jsonlines: http://jsonlines.org/)

made all filesystem interactions async and cleaned up any outstanding
issues with limit

with this added a new parameter -limit to dictate the maximum number of
points that can be handled when writing to a JSON file and also added
the -maxFilesize which dictates the maximum size in bytes of a JSON
file that we can handle when writing.

In order to avoid out of memory decided on using a -flushFrequency
option to specify how many points to wait before flushing and opting
for not doing the flushing based on time since 1) the code was getting
way more complicated to manage the constant async callbacks flushing
along with the final flush from eof() 2) its odd to keep doing async
flushes every N seconds when nothing is there to flush and having all
the gaurds in place to avoid reading/writing when there is no data to
actually flush.