test: rework the unit tests for the parsers to be more robust
Update the various parser tests to return a promise from each test
case and include the checks as part of the promise chain as opposed
to the previous approach in which the tests would throw expectation
errors from within the context of the emit callback.

This should help localize failures more clearly.

As part of this change, update the tests to properly handle the case
where the parser is given a limit on the number of points per emit
so that it properly expects an empty array at the end of the test.

This was causing intermittent failures in travis.