parser: Also collect ASTs in `parseSync`
In `parse`, the resulting set of modules is built by collecting their
ASTs in a closure variable (`asts`). In `parseSync`, it was built
recursively, but a closure variable (`names`) was needed anyway to track
circular dependencies.

This commit changes `parseSync` to also collect module ASTs in a closure
variable, unifying the approaches.