client-sink-mgr: pass the list of sinks to each view's constructor
Views may want to change their behavior based on which other views
are present in the program, so pass the full list of views into the
constructor when creating them.

Previously the client-sink-mgr "env" was passed to the constructor,
but none of the views used it, so that is no longer included.