table: fix progressive mode option to have no default and be truthy
Now that client-sink-mgr sets the progressive mode option, remove the
default from the view itself.

Also fix a problem where the view was inconsistent in how it checked
the option -- in one case it was checking for "truthiness" while in
the other it was checking for explicit "false", which caused problems
in the tests where the option was undefined.

When using progressive mode with more than one view in the program,
the output is interleaved and becomes basically unreadable. Change
the initialization so that progressive mode defaults to true only
if there is exactly one view in the program.