Rename client sink to view or client view.
We're trying to standardize on the name "view" for the endpoint of a
program that is viewed (either in the cli as a table/text block, or in
the browser). So go through a big rename exercise to clean that up.

Generally anywhere where the term sink was used when the thing was
actually a viewable item, switch to view. The term sink still has
meaning for the union of views and writes.

This change is incompatible with 0.1.0, so once published we'll have
to bump the version to 0.2.0.