Fix multiline inputs and history.
This fix includes 2 changes that both address the problem in different
ways. I think I'll keep both fixes just to allow for flexibility in
case we want to change the behavior later.

The first fix involves collapsing multi-line programs to a single line
when saving to history. That way, scrolling through history doesn't
result in an ugly multi-line item, and also allows for editing of the
item. When the multi-line program is run for the first time, however,
it's still run as a multi-line program. That allows any errors to be
shown in the context of the original multi-line program.

The second fix is to allow multiline inputs to work with history. The
problem was that the '.*' regex wasn't matching multiple lines
separated by newlines. Javascript regex apparently doesn't have a
multi-line option when compiling the regex, so just look for newlines
explicitly.