adapters: delay loading the adapter until it is used
Instead of eagerly loading all of the adapters when the runtime is
initialized, stash the configuration and defer loading the module
until the adapter is actually used.

This avoids irrelevant errors in loading adapters that aren't needed for the
given program and speeds up initialization in cases where not all of the
configured adapters are used.

This required a change to how the adapters are configured -- instead of
specifying the module to load as the key in the config file, it now expects the
name of the adapter, i.e. `elastic` instead of `juttle-elastic-adapter`.

By default, for an adapter named `xyz`, the system tries to load
`juttle-xyz-adapter`, though the path can still be overridden in the
adapter config.

Also renamed `adapters.load` to `adapters.configure` to better match the
semantics of what it actually does.

Fixes #83.