adapters: add location info to invalid adapter errors
Add support for passing location information to `adapters.get` so that errors
which occur when instantiating an adapter that isn't registered or configured
can contain location info.

As part of this change, modify the ast for read and write in order to preserve
the location information for the adapter name.

Relates to #83