adapters: delay initialization of the builtin adapters
Register the builtin adapters as part of adapters.configure().

This way they are dynamically loaded when they are used as opposed to being
registered as a side-effect of something loading in the juttle runtime.

In addition to unifying the code between external and builtin adapters, the
main benefit of this change is that the CLI adapter implementations wouldn't be
included in a browserify build of the runtime.

Note that part of this change renamed `stochastic-adapter.js` to `stochastic.js`
so that the builtin adapters use the same module naming conventions.