Lexicographic ordering of arrays
Define lexicographic ordering of arrays

Assume that the LHS and RHS are of the same 'shape'. Comparing
different data types at any level of nesting results in an error.

For arrays, define the lexicographic ordering naturally, shorter
arrays are smaller than longer ones, and the elements are compared by a
recursive call in the order of their indices.

For primitive data types (integers, strings, JuttleMoments, etc...),
fall back to the existing implementation.

Order on objects is not defined and will throw an error.