Remove `Program#get_stats`
The method is dead code. I don't think it makes sense to preserve it, it
is not a very good API anyway (I don't like the idea of carrying pnames
over to the runtime, I think stats for given node should be obtained by
getting a reference to that node directly and querying the `stats`
property).