adapters: avoid loading multiple copies of juttle
Because adapters pull in the Juttle runtime via require('juttle/lib/xyz'),
there's a chance that loading an adapter may load another copy of the runtime
code, depending on the specifics of how the user installed the various node
modules.

This causes problems if identical code is loaded from two different paths,
since `instanceof` will fail to match an object created from one copy to the
class definition in another copy.

To avoid these issues, before loading the adapter, intercept the node.js
module loader and if there is a conflict for juttle, log an informative
error and exit the process.

If the environment variable JUTTLE_RESOLVE_ADAPTER_CONFLICT exists, then
log a warning and rewrite any require requests for `juttle/*` to be relative
paths using the _current_ module (i.e.  lib/runtime/adapters.js) as the
context.

This ensures that we don't have strange errors when multiple copies of the
runtime are loaded into the same process.

Fixes: #130