procs.spec.js: tweak a couple tests to be more resilient to timing
Previously a couple tests could fail if :now: ended up aligned exactly on a
second boundary since the emitter would not send out the same number of points
as it did when :now: was not an exact second.

Change the tests so that both start and end are quantized to the even second so
this isn't a problem.