source: add base class validation for -from, -to, and -last
Add support in the base class to handle the various time bounds
options:

* ensure they are the appropriate types
* verifying that -from isn't greater than -to
* disallowing combinations of -from or -to with -last

Also add support for setting defaults for from / to with an optional
constraint that one of them must be supplied.

Removed the scattered logic for the above from the adapters and from
the implementation of emit.

As a side-effect of this change, we now support `-last :x:` in both
`emit` and the file adapter.