strings: Fix RT-INVALID-TERM-TYPE message text
The RT-INVALID-TERM-TYPE message is thrown in cases like this:

    read elastic [1, 2, 3]

Here, the filter expression consists of a simple filter term [1, 2, 3].
Since this term is neither a string (indicating a full-text search) nor
a value containing a filter ("first-class filter expression"), an error
is produced.

The message of the error assumed that the only valid use of simple terms
is full-text search. Since this is not the case, this commit fixes the
message to be more general.