Change FTS sytnax to use the ?= operator
Before this commit, FTS was expressed by using standalone strings in
filter expressions:

    read elastic "keyword"

This commit changes the syntax to use a new ?= operator:

    read elastic ?= "keyword"

The change includes updating tests and documentation.

The main purpose of introduing the new operator is to make FTS a bit
more explicit and to allow extension of the syntax to specific fields in
the future.

The introduction of the new operator required splitting existing
SimpleFilterTerm AST node into two new nodes: FulltextFilterTerm and
ReferenceFilterTerm. In the old world, the same syntax was used to
express FTS and references to values containing a filter ("first-class
filter expressions"), but in the new world, the syntax is different.
This means that FilterChecker needs to differentiate between these
situations to correctly check the filter expression, hence the need for
two distinct AST nodes.

Resolves #56.