procs: add read and write procs and adapter base classes
Add a regular proc implementation of read and write which is
instantiated by the graph compilier just like any other proc is.

Also add adapter base classes for read and write that do not
depend on the details of the proc implementation.

Currently both read and write first load the adapter and check if it is an
implementation of a source or sink proc. If so, it instantiates the legacy
adapter and shims the proc methods so that adapters that haven't yet been
converted will continue to work. This logic will be removed once all adapters
are converted.