http: port the http adapter to the new API
This turned out to be trivial for write but fairly invasive
for read.

Since the adapter itself no longer calls emit but instead returns
the points in the promise, it now has to implement an iterator
abstraction to buffer the points and then return them when the read
proc wants to get them back.