adapters: rework timeRequired to be a static adapter property
Update the adapter implementation contract to require a static `timeRequired`
property on the read class instead of setting a `timedSource` instance property
through the base class constructor.

This allows the `read` proc to determine the expected time option behavior
before instantiating the adapter itself, which ensures that a valid `from` and
`to` can always be given to the proc implementation as part of the options in
the constructor.