removed limitations from file adapter on write
By using the `jsoner` package we can append to JSON objects in a file
without having to load the whole file. This removes the need for those
various options to handle maximum number of points and maximum filesize.

Added a new `-append` flag which would default to not append by default.

updated documentation accordingly