filters: Don't semantically check FilterLiterals in FilterChecker
Inputs of type "filter" produce FilterLiteral AST nodes that need to be
semantically checked before they are compiled as part of a filter. The
check is needed mainly to convert any Variable nodes into field
references (UnaryExpression nodes with "*" as the operator).

Until now, these checks were done in FilterChecker after FilterLiteral
nodes coming from input values were embedded into filter expressions
used by procs in the build phase. This was not ideal for two reasons:

  1. FilterChecker is not supposed to modify checked filter AST, but
     applying a semantic check to FilterLiterals did modify it.

  2. We want to check the FilterLiteral itself, not its (possible many)
     embeddings.

This commit addresses the problems by moving the semantic check to
GraphBuilder and applying it to ASTs of all input values for generality.
This seems a little bit cleaner.