filter-js-compiler.spec.js: Do not use ASTs
Instead of specifying filters to compile using ASTs, specify them as
strings and parse them using parser.parseFilter. This makes tests
shorter, more readable, and less fragile.

Note that when specifying filters as strings we can't test handling of
negative InfinityLiteral (because "-Infinity" is parsed as a unary
operator and a positive InfinityLiteral) and FilterLiteral (because
there is no filter literal syntax). I consider this a small cost
compared to the overall improvement.