emit: rework to use ticks and enqueue from the source base
Remove the custom tick logic from emit to use the `emitTicks` logic
from the source base class.

As part of this change, fix up how emit actually handles ticks so that it
properly emits the ticks for every gap of more than 1 second between points
instead of waiting until the next point is ready to emit and only then
generating the backfilled ticks.

This had the side effect of changing the behavior to only generate ticks if
there is a gap in the points.