Don't call read more frequently than -every.
I noticed that with read XXX -every :10s:, the adapter's read method
was called once per second after an initial delay of 10 seconds. Mike
narrowed it down to here, where to was being advanced by the min of
the tick interval and -every value, when it should just be the -every
value.