procs.js - simpler channel name allocation
There is no need to maintain an object with channel names - a simple
integer counter will suffice.

The con is that if there are multiple types of channel names (e.g view
and test), each set won't have continuous range. But as there is only
view at the moment, this should be OK.