read: rework the way that sources specify their time behavior
The previous model in which adapters would convey statically whether
or not they should behave as a timed source or an untimed source
proved too limiting in practice, since it didn't lend itself well
to cases like influx or sql in which the behavior actually depends
on the other options given to the read call.

Instead, change things to give adapters more control over their
desired behavior for time options by adding two new API hooks,
defaultTimeRange() to return defaults for from and to, and
periodicLiveRead() which indicates to the read proc whether to do
the pseudo-live pattern for reading or leave it up to the adapter
to manage its own paging through time.