add append option for `write file`
fixes #286

This will enable the ability to append on `write file` for `jsonl` and
`csv` formats. We don't verify that the existing files are valid in any
way and the only "trick" here is that for CSV if the file exists we
assume it has an existing header with valid data and append to it.

documentation updated