Don't set Juttle.procs
In compiled Juttle programs, refer to procs using juttle.procs.*. This
is possible because they procs are added to the runtime. The change
makes the programs less dependent on the Juttle variable.

In other code, replace references to Juttle.procs with direct references
to individual procs or to the procs module.

Note the latter change created a circular reference between source.js
and base.js. Rather than working around it I decided to remove the
affected code (it was merely an assertion).