Represent field references in simpler way in the AST
Until now, field references in expressions like "a < 5" or "#b > 6" were
represented as UnaryExpression nodes with "*" as an operator and a
StringLiteral with the field name as an expression. The idea behind this
was that they are a syntax sugar for field dereference (the canonical
form being "*'a' < 5" or "*'b' > 6").

The UnaryExpression representation proved cumbersome to handle,
especially in adapter code dealing with static filter expressions.
Therefore, this commit gives field references more first-class status,
representing them as Field nodes.

All relevant places are updated to handle the new nodes (the code is
usually taken from UnaryExpression handling).

Note that changes done in StaticFilterChecker mean that expressions like
"*'a' < 5" are no longer considered valid in static filter expressions.
I see this as a fix of an irregularity in the language -- the "*"
operator is dynamic by its nature and it shouldn't be allowed in static
filter expressions, even with compile-time constant argument.

Fixes #318.