events: replace backbone.Events with eventemitter3
Go through the handful of places where we were using Backbone.Events
in juttle and replace with eventemitter3.

This required changing all calls from `events.trigger` to `events.emit`.
Note that the procs still contain a method `this.trigger` to send the
event as compared to `this.emit` which emits points.

Also instead of attaching the event emitter functions to the program
object that's shared between the procs, attach a handle on the event
emitter itself, changing all callers from `program.trigger` to
`program.events.emit`.