Move Juttle.extend_options to the compiler
Juttle.extend_options was a helper function responsible for processing
proc options -- specifically expanding nested options and -o/-options.
So far it did its work in runtime but there is no good reason for that
(options are known in compile-time already).

This commit moves Juttle.extend_options into GraphCompiler, which runs
it in compile-time. Note the move required making it work on AST nodes
instead of Juttle values.

The change makes compiled Juttle programs less dependent on the Juttle
variable.