fanin: Handle null time field in incoming points
Change a test probing for presence of the time field in a point so that
a null value is treated as if the time field was not set. This is in
line with other similar checks in the codebase. The end result is that
points with null time don't cause an exception.

Fixes #351.