http_server: port the http server adapter to the new API
Port the http server adapter to the new adapter API. The adapter now
buffers points that come in from the wire and expects read() to be
called periodically. If points exist in the buffer, they are returned
immediately, otherwise it a promise with a timeout to wait for up to
the `-every` ms for points to arrive.

This required reworking the tests a bit to properly wait for all
the timers to fire before checking the program results.