procs: change procName property to a function
As part of the cleanup to prepare to use ES6 classes, change procName
from being a prototype property to being a function.

By default this just returns the name of the proc, but read and write
override the implementation to include the name of the adapter.

In order for this to work during the the execution of the base class
constructor to initialize the logger, modify the base class to stash the params
in this.params alongside the options.