parser: Create StringLiteral with correct property
When parsing expressions like "f().x", the resulting MemberExpression
node contained a StringLiteral with the "name" property set instead of
correct "value". This commit fixes that.

Interestingly, expressions like the above don't compile anyway because
of incomplete handling of MemberExpression in the semantic pass and
elsewhere in the compiler. I plan to fix that later.