ast: Kill ReducerCall
ReducerCall nodes were created by the semantic pass to distinguish
function and reducer calls. However, this is not needed because they are
already distinguished by the value of the "name.symbol.type" property
(which is either "function" or "reducer").

This commit eliminates ReducerCall and sets "name.symbol.type" in two
cases where it wasn't set in the semantic pass.