emit.js - use enqueue emitting directly
As emit is the only source working with named outputs, move the
original version of enqueue that supports them to emit and update the
parent class' implementation to just fallback to 'default' output.

The named outputs will be removed later and emit will need to be updated
to either use the scheduler or keep its own point emitting logic.