scope: Rename the "name" property of symbols to "uname"
The "name" property of symbols stores a string (usually an identifier)
that is emitted into the compiled code when the symbol is referenced.
This concept is called "uname" (unique name) elsewhere in the compiler.

This commit renames the "name" property of symbols to "uname" to align
the terminology and prevent confusion with Juttle-level symbol name.