Do not coerce identifiers in multi-value options to strings
Before this commit, option values were parsed using the ByList rule.
This was an easy way to parse multi-value options, but it meant the
parser returned multiple values as ByList nodes, which triggered
coercion rules for "by" clauses in the semantic pass. This was
undesirable.

This commit changes the grammar to parse options using a new OptionValue
rule which returns multiple option values as ArrayLiteral nodes. These
don't trigger coercion in the semantic pass.

Note the change also means options no longer auto-flatten their values
(this was another side effect of using ByList). I think this is OK --
the behavior was surprising and it probably needs some tweaking even in
case of "by" clauses.

Fixes #444.