error: add toJSON so that message is included when serializing
As a side-effect of changing from extendable-base to ES6 classes, and
due to the fact that the `message` property is not enumerable when
passed to the Error constructor, the message wasn't actually included
in the JSON representation of an Error.

To fix this, add a toJSON implementation that explicitly enumerates
message, code, and info. This does not include the name of the error
class, which makes things compatible with how they were in juttle 0.4
and earlier.