semantic.js - disallow empty subs
Previously, defining an empty sub resulted in the following error from
the compiler:

```
$ bin/juttle -e 'sub s() {}; s'
Error: Cannot read property 'head' of undefined
```

As the sub didn't contain any nodes, it didn't return any which the
compiler could attach into the surrounding flowgraph.

A simple fix is to regard empty subs as an error. This is consistent
with handling of an empty program / flowgraph.

Fixes: #433