graph-compiler.js - pass program as an argument in compiler
Procs access a `program` object and its properties, some of which
(assingments for `channels` and `now`) are generated by the compiler.

After the code produced by the compiler is evaluated, additional
properties (`scheduler`) are assigned to `program`.

And as a part of `activate` step, the `program` object is further
extended with `env` and `events` properties.

Instead of generating `program` in the compiler, then extending it in
various ways, pass instance of `Program` class to the compiler and
ensure that all properties required by procs are added during
`constructor` call.

This commit passes in the instance. The previous commits ensured that
the `Program` instance has all the properties, subsequent commits will
remove the now obsolete code.