location-stripper: make it an option in the compiler
Instead of stripping locations from the AST only after the compiler
is completed, add a flag to the compiler to strip the locations right
after parsing.

This way locations can be stripped from all stages of the compiler,
including the parse tree, the flowgraph, as well as the build / compile
output.

When viewing the compilation output for these latter stages, the
generated code does contain `location: undefined`, but this is better
than containing the full location information by default.

Finally, change the location-stripper module to export a function
instead of a class to make it simpler to use at the call site.