compiler: Do not use node.computed when compiling MemberExpression
So far, the idea was that non-computed member expressions ("foo.bar")
were always static, with the base ("foo") always referring to a module,
and that computed member expressions ("foo['bar']") were always dynamic,
with the base referring to a Juttle value. This was reflected in the
compiler which compiled these member expressions differently, telling
them apart using the "computed" property.

This idea will get disrupted soon because we also want to enable dynamic
non-computed member expressions. This means discriminating using the
"computed" property won't work. Therefore, this commit adjusts the
compiler to use the "symbol" property instead. Right now, this works
identically (because we set "symbol" exactly for non-computed member
expressions), but the idea is that in the future only static member
expressions will reference a symbol for an entity exported from a module
they refer to while dynamic expressions (including dynamic non-computed
expressions) won't.