README: update to make juttle-engine more prominent
Since the more complete installation of juttle will want to use
juttle-engine and not just the CLI, change the introduction and
installation instructions to refer to both options.

Also remove the section talking about juttle engine since we now
just link over to that repository. A corresponding update to that
README will come shortly.