SemanticPass: Change info stored in this.reducers
Instead of storing the last component of reducer name and optionally its
module name in this.reducers, store its full name and symbol. This means
that now redundant symbol lookup in sa_reducers can be removed and that
two this.reducers.push calls are more similar, which will enable future
refactoring.