Add more detail on how to write adapters.
Take most of the general discussion from the gmail adapter docs at
https://github.com/juttle/juttle-gmail-adapter/blob/master/docs/adapter_impl_notes.md
and add them to this doc. Specifically, added the following:

 - How juttle loads an adapter
 - General discussion of live/historical and how that relates to the data read/written by an adapter.
 - Notes on configuration.
 - Options supported by many read commands.
 - Notes on the filter expression and how its parsed version is made available to the adapter.
 - Notes on write and expectations for write (don't block event loop) and eof (return a promise that resolves when everything is written).
 - Logging
 - Returning errors.

Once this is merged I'll change the gmail adapter docs to take out the
general stuff and tailor it more closely to the specific
choices/options used by the gmail adapter.