Add roadmap image and ecosystem document.
This is based on feedback from Brian Colby, who was using Juttle for a
while and then mentioned that he missed the browser-based
visualizations, not knowing that juttle-engine existed.

Add a roadmap image to the top README that shows where juttle lies in
the overall ecosystem. The roadmap image links to a new ecosystem page
that lists all the main projects and a larger ecosystem image with
everything labeled.