Star operator ambiguity tests use test adapter
One of the tests was relying on stochastic adapter emitting a certain
number of points and was failing non-deterministically.

Either the bug is in reduce, or in stochastic adapter (which is known to
sometimes emit ooo points that are automatically dropped).

From the syntax POV, an adapter is an easy place to trigger the
ambiguity which we want to test, i.e. to get a number/duration literal
nearby a field dereference.

Change the adapter from stochastic to test and modify the irrelevant
test condition that was counting points to be stable.

Fixes: #525