juttle-test-utils: by default deactivate after the program is done
The change in #522 called deactivate immediately after running the
program unless deactivateAfter was set, which is the wrong semantics
since it will cause the programs to terminate early.

Instead, when deactivateAfter isn't set, wait until the program is
done before calling deactivate.

Also fix the call to Promise.delay -- the delay time is supposed to
be the only argument in bluebird promises.