SemanticPass: Move variable checks from Build and GraphCompiler
Build and GraphCompiler checked that entities such as module references
or functions were not used as variables. This check was semantic in
nature so this commit moves it to the semantic pass.

Note the check became more complicated by the move. This is because in
SemanticPass, one can get to sa_Variable either by processing a "bare"
variable (one that has no ".", "[...]" or "(...)" operator applied to
it) or a variable inside a member expression (e.g. "foo" in "foo.bar" or
"foo['bar']"). But in gen_Variable in Build or GraphCompiler, one
couldn't get to gen_Variable when the variable was a module reference
inside a member expression, which simplified things a bit.

This change will enable further refactoring of sa_MemberExpression.