SemanticPass: Generalize the "in_member_expression" flag
Generalize the "in_member_expression" flag (a boolean) into "expr_mode",
(a string). Current allowed values are:

  * "result" -- equivalent of in_member_expression == false
  * "member" -- equivalent of in_member_expression == true

In the future there will be be also a "call" value, which is the reason
for this refactoring. The refactoring itself is done a bit earlier than
necessary in order to save rewriting code when the flag starts to be
used more widely (like, in the next commit).