SemanticPass: Fix moved variable checks
Variable checks in the semantic pass were incomplete. For example, the
compiler correctly reported that "f" can't be used as a variable in
cases like this:

    function f() {
    }

    const c = f[0];

but not in cases like this:

    function f() {
    }

    const c = (f + 1)[0];

This was because expr_mode wasn't reset back to "result" when processing
expressions deeper in MemberExpression's object.

This commit adds necessary resets to methods corresponding to all
expressions that can be processed in "member" mode, which fixes the
problem.