SemanticPass: Do not specify entity in NOT-EXPORTED exceptions
The NOT-EXPORTED exception specified expected entity type in its error
message. This was possible to do because processing of variables,
functions, reducers, and subs in the semantic pass was separate. Soon,
this won't be true and the processing will be unified, which means
specifying expected entity type won't be practical. As a result, this
commit removes the expected entity type from the NOT-EXPORTED exception,
which makes it similar to UNDEFINED.

Note the expected type was wrong in case of consts (the message talked
about variables).