Allow ctrl-c to stop programs (politely)
Allow ctrl-c to stop programs by adding a signal handler for SIGINT
that calls program.deactivate(). bin/juttle must also set standard
input to non-raw mode while the program is running so ctrl-c isn't
read by the cli directly.

Read already handles teardown properly in a limited way--after any
pending read has completed, the adapter is torn down and emits an
eof. An in-progress read is not stopped, though.

Emit required changes, though--add a teardown method that sets
a flag. The next time run() is called, it checks the flag and emits
eof() if set.

With these changes, a program won't be stopped until the next emit
happens or a read in progress has completed. If we want immediate
program cancellation, we'd have to make some additional changes:

 - Modify the scheduler to return ids from all calls to schedule so
   any pending callbacks could be cancelled, or do a wholesale
   teardown itself.

 - Start cancelling the read promise in read's teardown.

I think the current polite teardown is fairly nice as it results in a
relatively clean set of data and stil mostly meets the goal of
allowing program teardown, so I think we should try it first.