SemanticPass: Change d-bit computation in sa_MemberExpression
For MemberExpressions representing module access, the d-bit was always
set to true in sa_MemberExpression (because both node.object.d and
node.property.d were always true). This was OK when sa_MemberExpression
was used only for module consts, but soon it will be used also for
module functions and reducers. In that case, the d-bit shouldn't be set
mechanically to true, because MemberExpression node is only part of the
overall expression and it is not compilable in build time by itself.
Setting it would confuse the compiler.

This commit changes d-bit computation for MemberExpressions representing
module access to take the d-bit from the symbol, resolving the issue.